backdoors-design/requirements.txt
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
netifaces==0.10.9
paramiko==2.7.2
pycparser==2.20
pycryptodome==3.9.9
PyNaCl==1.4.0
requests==2.25.0
scapy==2.4.4
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
